# openapi-ci-api.yaml
openapi: 3.0.3
info:
  title: CCExtractor CI System API
  version: 1.2.0
  description: |
    Security-hardened JSON-only REST API for the CCExtractor CI/sample platform.
    Designed for AI agents and CI automation. Enforces scoped Bearer token auth,
    strict input validation, rate limiting on all routes, and safe defaults
    throughout. No browser sessions, no HTML, no implicit permissions.

    **Authentication:** All endpoints require bearer token authentication unless
    explicitly marked with `security: []` (only /system/health and POST /auth/tokens).

    **Rate-limit headers:** Every response includes `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` headers. These are modeled
    explicitly on the 429 response for brevity; they are present on all responses
    regardless of status code.

  contact:
    name: CCExtractor Development
    url: https://github.com/CCExtractor/sample-platform
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://sampleplatform.ccextractor.org/api/v1
    description: Production

#
# Global security: all endpoints require auth
# unless explicitly overridden with security: []
#
security:
  - bearerAuth: []
...
paths:
...
  /runs/{run_id}/errors:
    get:
      tags: [Errors and Logs]
      summary: Get structured test errors for a run
      operationId: listRunErrors
      description: >
        Error types are derived from TestResult and TestResultFile rows.
        missing_output is detected from the dummy (-1,-1,-1,'','error') row
        pattern, not from got=null (which means match, not missing).
      security:
        - bearerAuth: []
      x-required-scope: results:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: type
          in: query
          schema:
            type: string
            enum: [test_failure, exit_code_mismatch, missing_output, diff_mismatch]
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, error, critical]
        - name: sample_id
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Paginated test errors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ErrorItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/infrastructure-errors:
    get:
      tags: [Errors and Logs]
      summary: Get worker, provisioning, and build errors for a run
      operationId: listInfraErrors
      description: >
        Errors are extracted from TestProgress rows written by the CI worker.
        Messages are currently unstructured text. The type filter does
        best-effort text matching until the worker protocol emits structured
        error types.
        Stack traces are opt-in (include_stack defaults to false) to avoid
        leaking internal paths to unauthorized callers.
      security:
        - bearerAuth: []
      x-required-scope: system:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: type
          in: query
          schema:
            type: string
            enum: [queue, vm_provisioning, checkout, merge, build, worker, web_server, storage]
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, error, critical]
        - name: include_stack
          in: query
          schema:
            type: boolean
            default: false
          description: >
            Default false. Set true only when debugging infrastructure failures.
            Stacks may contain internal paths; access requires system:read scope.
      responses:
        "200":
          description: Paginated infrastructure errors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ErrorItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/logs:
    get:
      tags: [Errors and Logs]
      summary: Get raw logs for a run
      operationId: getRunLogs
      description: >
        Logs are stored at SAMPLE_REPOSITORY/LogFiles/{id}.txt and served
        via GCS signed URL. Returns 404 — not a broken download link — when
        the file is absent from both local and GCS storage.
        Uses cursor-based pagination.
      security:
        - bearerAuth: []
      x-required-scope: system:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warning, error, critical]
        - name: source
          in: query
          schema:
            type: string
            enum: [orchestrator, worker, build, test_runner, web]
        - name: contains
          in: query
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Cursor-paginated run log lines
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/CursorPage"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/LogLine"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Log file not found in local or GCS storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: log_not_found
                message: Log file for run 9309 does not exist in any storage backend.
                details:
                  run_id: 9309
                  checked: [local, gcs]
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/samples/{sample_id}/logs:
    get:
      tags: [Errors and Logs]
      summary: Get raw logs for a regression test result in a run
      operationId: getSampleLogs
      description: >
        Returns raw log lines for a specific regression test result.
        Logs are stored at SAMPLE_REPOSITORY/LogFiles/ and served via GCS
        signed URL when available. Returns 404 when the log file is absent
        from both local and GCS storage.
      security:
        - bearerAuth: []
      x-required-scope: system:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/SampleId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warning, error, critical]
        - name: contains
          in: query
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Cursor-paginated sample log lines
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/CursorPage"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/LogLine"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/error-summary:
    get:
      tags: [Errors and Logs]
      summary: Get grouped error summary for a run
      operationId: getErrorSummary
      description: >
        Use this endpoint to triage a run before drilling into individual
        errors. group_by=type gives a high-level failure breakdown;
        group_by=sample_id helps identify flaky samples.
      security:
        - bearerAuth: []
      x-required-scope: results:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: group_by
          in: query
          schema:
            type: string
            enum: [type, sample_id, regression_id, severity]
            default: type
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, error, critical]
      responses:
        "200":
          description: Paginated grouped error summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ErrorSummaryBucket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"

  # SYSTEM
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Opaque server-side API token. Obtain via POST /auth/tokens.
        The CI worker token used by /ci/progress-reporter is a separate
        secret and is NOT valid here. Never use browser session cookies
        for API clients.

  # HEADERS
  headers:
    RateLimitLimit:
      description: Maximum requests allowed in the current window
      schema:
        type: integer
        example: 120
    RateLimitRemaining:
      description: Requests remaining in the current window
      schema:
        type: integer
        example: 117
    RateLimitReset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        example: 1748908800

  # PARAMETERS
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return (1–100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of results to skip for pagination
      schema:
        type: integer
        minimum: 0
        maximum: 2147483647
        default: 0
    Cursor:
      name: cursor
      in: query
      description: >
        Numeric line offset or ID for cursor-based pagination. Do not mix with offset. Mixing cursor and offset returns 400.
        Obtain next_cursor from the previous response's pagination object.
      schema:
        type: integer
        minimum: 0
        maximum: 10000000
    RunId:
      name: run_id
      in: path
      required: true
      description: Numeric run ID
      schema:
        type: integer
        minimum: 1
    SampleId:
      name: sample_id
      in: path
      required: true
      description: Numeric media sample ID
      schema:
        type: integer
        minimum: 1
...
  responses:
    BadRequest:
      description: Request body or query parameters failed schema validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: validation_error
            message: Request failed schema validation.
            details:
              fields:
                commit_sha: Must match pattern ^[a-fA-F0-9]{40}$
                platform: Must be one of [linux, windows]
    Unauthorized:
      description: Missing, expired, or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: unauthorized
            message: Bearer token is missing, expired, or invalid.
            details: {}
    Forbidden:
      description: Token is valid but lacks the required scope or role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: forbidden
            message: Token does not have the required scope for this operation.
            details:
              required_scope: runs:write
              token_scopes: [runs:read, results:read]
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: not_found
            message: Run 9317 not found.
            details:
              resource: run
              id: 9317
...
    RateLimited:
      description: Too many requests. Retry after the indicated number of seconds.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 30
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: rate_limited
            message: Rate limit exceeded. Retry after 30 seconds.
            details:
              retry_after: 30
              limit: 120
              window: 60s
    Error:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # SCHEMAS
  schemas:
    Page:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          description: >
            Result items. The concrete type is defined by allOf composition
            in each endpoint response.
          items: {}
        pagination:
          type: object
          required: [limit, offset, total]
          properties:
            limit:
              type: integer
              minimum: 1
            offset:
              type: integer
              minimum: 0
            total:
              type: integer
              minimum: -1
              nullable: true
              description: >
                Total matching records. Null if count was not computed for this request.
                Pass ?count=true to force computation.
            next_offset:
              type: integer
              minimum: 0
              nullable: true
            truncated:
              type: boolean
              description: >
                Present and true when the result set was capped by an
                internal safety limit (e.g. status-filter on runs). When
                true, total may undercount the real number of matches.
    CursorPage:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          description: >
            Result items. The concrete type is defined by allOf composition
            in each endpoint response.
          items: {}
        pagination:
          type: object
          required: [limit, next_cursor]
          properties:
            limit:
              type: integer
              minimum: 1
            next_cursor:
              type: integer
              minimum: 0
              nullable: true
              description: >
                Numeric cursor for the next page. Null when there are no
                more results.
    ErrorResponse:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          maxLength: 100
          description: Machine-readable error code (snake_case)
          example: not_found
        message:
          type: string
          maxLength: 500
          description: Human-readable error summary
          example: Run 9317 not found.
        details:
          type: object
          additionalProperties: true
          description: >
            Structured context for the error. Always an object, never null.
            Empty object {} when no additional detail is available.
...
    ErrorItem:
      type: object
      required: [error_id, run_id, type, severity, message, occurred_at]
      properties:
        error_id:
          type: string
          maxLength: 100
        run_id:
          type: integer
          minimum: 1
        sample_id:
          type: integer
          minimum: 1
          nullable: true
        regression_id:
          type: integer
          minimum: 1
          nullable: true
        type:
          type: string
          enum: [test_failure, exit_code_mismatch, missing_output, diff_mismatch, queue, vm_provisioning, checkout, merge, build, worker, web_server, storage]
          maxLength: 100
        severity:
          type: string
          enum: [info, warning, error, critical]
        message:
          type: string
          maxLength: 1000
        location:
          type: object
          nullable: true
          additionalProperties: true
          properties:
            file:
              type: string
              maxLength: 500
              nullable: true
            line:
              type: integer
              minimum: 0
              nullable: true
            column:
              type: integer
              minimum: 0
              nullable: true
            sample_name:
              type: string
              maxLength: 255
              nullable: true
        stack:
          type: array
          maxItems: 50
          description: Only present when include_stack=true was requested.
          items:
            type: string
            maxLength: 2000
        occurred_at:
          type: string
          format: date-time
    LogLine:
      type: object
      required: [timestamp, level, source, message, run_id]
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [debug, info, warning, error, critical]
        source:
          type: string
          enum: [orchestrator, worker, build, test_runner, web]
        message:
          type: string
          maxLength: 4000
        run_id:
          type: integer
          minimum: 1
        sample_id:
          type: integer
          minimum: 1
          nullable: true
    ErrorSummaryBucket:
      type: object
      required: [key, count, severity, group_by]
      properties:
        group_by:
          type: string
          enum: [type, sample_id, regression_id, severity]
          description: The dimension this bucket is grouped by.
        key:
          type: string
          maxLength: 100
          description: >
            Value of the group_by dimension. When group_by=sample_id or
            regression_id, this is an integer serialized as a string.
        count:
          type: integer
          minimum: 0
        severity:
          type: string
          enum: [info, warning, error, critical]
        sample_ids:
          type: array
          maxItems: 1000
          items:
            type: integer
            minimum: 1
        first_seen_at:
          type: string
          format: date-time
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
...